# public/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/photo/create": {
      "post": {
        "operationId": "PhotoController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePhotoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Photos"
        ]
      }
    },
    "/v1/photo/create-multiple": {
      "post": {
        "operationId": "PhotoController_createMultiple",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Photos"
        ]
      }
    },
...
    "/v1/photo/update": {
      "put": {
        "operationId": "PhotoController_updatePhoto",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePhotoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Photos"
        ]
      }
    },
    "/v1/photo/delete": {
      "delete": {
        "operationId": "PhotoController_deletePhoto",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeletePhotoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Photos"
        ]
      }
    },
    "/v1/photo/delete-multiple": {
      "delete": {
        "operationId": "PhotoController_deleteMultiplePhotos",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Photos"
        ]
      }
    },
    "/v1/photo/{id}": {
      "get": {
        "operationId": "PhotoController_getPhotoById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The unique identifier of the photo",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Photos"
        ]
      }
    },
    "/v1/photo/album/{albumId}/user": {
      "get": {
        "operationId": "PhotoController_getPhotosByAlbumIdAndUserId",
        "parameters": [
          {
            "name": "albumId",
            "required": true,
            "in": "path",
            "description": "The ID of the album which to get the photos",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Photos"
        ]
      }
    },
    "/v1/photo/albums/user": {
      "post": {
        "operationId": "PhotoController_getPhotosByAlbumIdsAndUserId",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetMultipleAlbumPhotoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Photos"
        ]
      }
    },
...
    "/messaging/send/{username}": {
      "post": {
        "operationId": "MessagingController_sendMessage",
        "parameters": [
          {
            "name": "username",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Messaging"
        ]
      }
    },
...
    "/messaging/messages/{conversationUserId}": {
      "get": {
        "operationId": "MessagingController_getMessages",
        "parameters": [
          {
            "name": "conversationUserId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Messaging"
        ]
      }
    },
    "/messaging/searchUsers": {
      "get": {
        "operationId": "MessagingController_searchUsers",
        "parameters": [
          {
            "name": "query",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Messaging"
        ]
      }
    },
...
    "/messaging/delete/{messageId}": {
      "delete": {
        "operationId": "MessagingController_deleteMessage",
        "parameters": [
          {
            "name": "messageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Messaging"
        ]
      }
    },
    "/messaging/deleteConversation/{conversationUserId}": {
      "delete": {
        "operationId": "MessagingController_deleteConversation",
        "parameters": [
          {
            "name": "conversationUserId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Messaging"
        ]
      }
    },
    "/messaging/userInfo/{username}": {
      "get": {
        "operationId": "MessagingController_getUserInfo",
        "parameters": [
          {
            "name": "username",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Messaging"
        ]
      }
    },
...
  },
  "info": {
    "title": "XENAuth API",
    "description": "The official XENAuth API documentation",
    "version": "1.0",
    "contact": {
      "name": "Erzen Krasniqi",
      "url": "https://erzen.tk",
      "email": "erzenkrasniqi@matrics.io"
    },
    "termsOfService": "https://erzen.tk/terms",
    "license": {
      "name": "AGPL-3.0",
      "url": "https://www.gnu.org/licenses/agpl-3.0.en.html"
    }
  },
...
  "servers": [
    {
      "url": "https://localhost:3000",
      "description": "LOCAL server"
    },
    {
      "url": "https://api.erzen.xyz",
      "description": "DEV server"
    },
    {
      "url": "https://apis.erzen.xyz",
      "description": "PRODUCTION server"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreatePhotoDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "description": "The URL of the photo",
            "example": "http://example.com/photo.jpg"
          },
          "caption": {
            "type": "string",
            "description": "The caption for the photo",
            "example": "A beautiful sunset"
          },
          "albumIds": {
            "description": "The IDs of the albums this photo belongs to",
            "example": [
              1,
              2,
              3
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "url"
        ]
      },
      "UpdatePhotoDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "The unique identifier of the photo"
          },
          "url": {
            "type": "string",
            "description": "The URL of the photo"
          },
          "caption": {
            "type": "string",
            "description": "The caption for the photo"
          }
        },
        "required": [
          "id",
          "url"
        ]
      },
      "DeletePhotoDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "ID of the photo to be deleted"
          }
        },
        "required": [
          "id"
        ]
      },
      "GetMultipleAlbumPhotoDto": {
        "type": "object",
        "properties": {
          "albumIds": {
            "description": "Array of album IDs to get the photos from",
            "type": "array",
            "items": {
              "type": "number"
            }
          }
        },
        "required": [
          "albumIds"
        ]
      },
...
      "MessageDto": {
        "type": "object",
        "properties": {}
      }
    }
  },
  "security": [
    {
      "access-token": []
    }
  ]
}